Added Swagger to Product Management System

+++ Great Outdoor/ProductManagementSystem/src/main/java/com/greatoutdoor/productmanagementsystem/service/ProductServiceImpl.java
@@ -50,6 +50,9 @@ public class ProductServiceImpl implements ProductService{
 
 	@Override
 	public boolean deleteProduct(String productId) {
+		if(productDao.findById(productId).isEmpty()) {
+			return false;
+		}
 		productDao.deleteById(productId);
 		return true;
 	}
